Created a basic expression language called MScript.
It supports 5 basic types: numbers, vectors (arrays of numbers), strings, functions and 'none' (the absence of a value).
Standard operators are provided: unary: -, !, binary: +, -, *, /, %, ==, !=, >, >=, <, <=, tertiary: ?;, if-else.
+ performs string concatenation if one of the operands is a string. The other arithmetic operators only operate on numbers and/or vectors.
== only returns true if both operands are of the same type, and have the same value.
The comparison operators only work with numbers.

Unlike null in most languages, 'none' can be used in most operations: it acts as the number 0 when the other operand is a number or vector,
and it acts as an empty string if the other operand is a string (except for equality checks: none is only equal to none).

Instead of having a separate boolean type, all types can be used in a boolean context: none, 0.0, [] and '' are false, anything else is true.

Vectors and strings can be indexed. Negative indexes start at the end of the vector/string (-1 being the last element).
Indexing returns none if the index is out-of-range.

Because vectors are used almost exclusively to represent positions, rotation angles and colors/brightness, vectors have x/y/z, pitch/yaw/roll,
r/g/b/brightness properties that act as shorthands for indexing with index 0, 1, 2 and 3 respectively (that is, x, pitch and r all return the first element).