macro_brush entities now also copy point entities from the selected template. Point entities are positioned relative to the chosen anchor point/offset.
Fixed that expressions in macro_brush attributes weren't evaluated, and that spawnflag and transform-related attributes weren't updated.
Also fixed that expressions in entities in the selected template were evaluated too early.
Note that the 'nth()' function no longer returns a different value for each entity in the selected template, because all those entities are part of a single instance.