Added support for .mtb (MESS Template Bundle) and .ted (Template Entity Definition) files, to make sharing template entities easier.
MESS will read all .ted files in the templates directory, and use their rewrite rules when processing maps. This means that users don't need tell MESS to use specific .fgd files.
MESS can also read maps and .ted files that are stored inside .mtb files (which are just zip files with a different extension),
so sharing a template entity is as easy as giving someone an .mtb file and telling them to place it in their MESS templates directory.