Rewrite directives can now be given a FOR clause that specifies which entity classname(s) to directive applies to, and a WHEN clause that specifies a scriptable condition.
Both clauses are optional. If both clauses are used together, then the directive will only be applied to entities that have the right classname and for which the condition evaluates to true.
Directives with clauses do not need to be followed by an entity definition, and if they are, they will not be associated with it.