Created another template entity: mtl_trigger_random. This template uses the env_beam/func_button trick for random target selection.
This entity provides two modes:
- single-target mode will trigger a random target each time the mtl_trigger_random is triggered.
- continuous mode will keep triggering random targets until the mtl_trigger_random is triggered again (toggled).

In continuous mode, the trigger can explicitly be enabled/disabled via trigger_relays (or +/- target patterns). The trigger_random also has a flag for auto-starting.
The buttons do not block the player, and it appears that the env_beam is unable to damage the player because it's so small, but it's probably still better to place this entity somewhere out of reach.