The multi-target pattern now also supports optional delays. Delays can be specified by adding a colon followed by a number to a target: "target1:delay1,target2:delay2".
If any delays greater than zero are present, then the generated multi_manager will be multi-threaded, to ensure that it can be triggered again at any time.
Multi-target patterns used within a multi_manager are merged into that multi_manager, and their delays are added to the delay of the multi-target entry.