Implemented saving maps to .jmf files (version 121 and 122).
Jmf files store quite a bit of internal editor state that isn't strictly required because the editor can derive it from entity properties and fgd settings.
There are also a few sections whose meaning is currently unknown, but these appear to simply be unused.
The current approach is to store all this data in Jmf-specific data types (JmfMap, JmfBrush, JmfEntity, etc.) that inherit
from the standard map data types (Map, Brush, Entity, etc.). This helps to minimize data loss when converting between different jmf file versions,
without breaking normal MESS map processing.

NOTE: Despite this, reading a jmf file and saving it will not always produce a file that is byte-for-byte the same. One reason is that texture name buffers
sometimes contain parts of previous texture names. When a texture is replaced, JACK only writes the new name and a null terminator to the buffer, so if the
previous texture name was longer then a part of it will remain. MESS ignores anything after the null terminator.
Another reason is that jmf files sometimes contain empty groups. MESS will discard these. These differences to not affect the contents of the map.

NOTE: The current approach does not synchronize Jmf-specific data when standard entity and brush properties are modified. It is only meant for conversions
between different jmf file versions!