Implemented saving maps to .rmf files (version 2.2).
Rmf files store a bit of editor state, such as groups and VIS groups, but not nearly as much as jmf files. The purpose of some data sections is still unknown,
but they appear to be unused. Entities and world brushes are interspersed, but the original order is retained by using Rmf-specific data types
(RmfBrush, RmfEntity, etc.) in order to minimize data loss when converting between different rmf file versions (support for versions older than 2.2 is yet to be added).

NOTE: Reading an rmf file and saving it will not always produce a file that is byte-for-byte the same. For example, cylinder faces are marked with a flag
which is not preserved (the purpose of this flag is unclear). Also, VIS group name buffers may contain random data past the null terminator - this data
is not preserved either. Another complication is that both Hammer and JACK can save .rmf files, and there are subtle differences between the two.
MESS uses Hammer's behavior as a guideline.